requirements.txt
opencv-python>=4.6.0,<4.7.0
numpy>=1.24.3,<1.25.0
Pillow>=9.4.0,<9.5.0
scikit-image>=0.19.3,<0.20.0